# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API REST Hoteles",
        "description": "API REST Hoteles",
        "contact": {
            "email": "cesarpozo25@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost/apirest-hoteles/public/",
    "basePath": "",
    "paths": {
...
        "/habitacions": {
            "get": {
                "summary": "Get Habitaciones",
                "parameters": [
                    {
                        "name": "Autorization",
                        "in": "header",
                        "description": "Bearer {token_access}",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#definitions/Habitacion"
                            }
                        }
                    },
                    "403": {
                        "description": "Autorization Exception",
                        "schema": {
                            "$ref": "#definitions/Errors403"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#definitions/Errors500"
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "summary": "Create Habitacion for store",
                "parameters": [
                    {
                        "name": "Autorization",
                        "in": "header",
                        "description": "Bearer {token_access}",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "numero": {
                                    "type": "string",
                                    "example": "120"
                                },
                                "Hotel_id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "tipo_habitacions": {
                                    "type": "integer",
                                    "example": 1
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Create successful operation",
                        "schema": {
                            "$ref": "#definitions/Habitacion"
                        }
                    },
                    "403": {
                        "description": "Autorization Exception",
                        "schema": {
                            "$ref": "#definitions/Errors403"
                        }
                    },
                    "404": {
                        "description": "Not Found Exception",
                        "schema": {
                            "$ref": "#definitions/Errors404"
                        }
                    },
                    "406": {
                        "description": "Not Aceptable",
                        "schema": {
                            "$ref": "#definitions/Errors406"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#definitions/Errors500"
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/habitacions/{habitacion_id}": {
            "get": {
                "summary": "Get Habitacion",
                "parameters": [
                    {
                        "name": "habitacion_id",
                        "in": "path",
                        "description": "un numero id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Autorization",
                        "in": "header",
                        "description": "Bearer {token_access}",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#definitions/Habitacion"
                            }
                        }
                    },
                    "403": {
                        "description": "Autorization Exception",
                        "schema": {
                            "$ref": "#definitions/Errors403"
                        }
                    },
                    "404": {
                        "description": "Not Found Exception",
                        "schema": {
                            "$ref": "#definitions/Errors404"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#definitions/Errors500"
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "put": {
                "summary": "Update Habitacion",
                "parameters": [
                    {
                        "name": "Autorization",
                        "in": "header",
                        "description": "Bearer {token_access}",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "habitacion_id",
                        "in": "path",
                        "description": "un numero id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "data",
                        "in": "body",
                        "required": false,
                        "schema": {
                            "properties": {
                                "numero": {
                                    "type": "string",
                                    "example": "120"
                                },
                                "Hotel_id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "tipo_habitacions": {
                                    "type": "integer",
                                    "example": 1
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create successful operation",
                        "schema": {
                            "$ref": "#definitions/Habitacion"
                        }
                    },
                    "403": {
                        "description": "Autorization Exception",
                        "schema": {
                            "$ref": "#definitions/Errors403"
                        }
                    },
                    "404": {
                        "description": "Not Found Exception",
                        "schema": {
                            "$ref": "#definitions/Errors404"
                        }
                    },
                    "406": {
                        "description": "Not Aceptable",
                        "schema": {
                            "$ref": "#definitions/Errors406"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#definitions/Errors500"
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete Habitacion",
                "parameters": [
                    {
                        "name": "habitacion_id",
                        "in": "path",
                        "description": "un numero id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Autorization",
                        "in": "header",
                        "description": "Bearer {token_access}",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#definitions/Habitacion"
                            }
                        }
                    },
                    "403": {
                        "description": "Autorization Exception",
                        "schema": {
                            "$ref": "#definitions/Errors403"
                        }
                    },
                    "404": {
                        "description": "Not Found Exception",
                        "schema": {
                            "$ref": "#definitions/Errors404"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#definitions/Errors500"
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
...
}